introduce doom emacs
Wow it's so much faster!

+++ emacs-install.sh
@@ -40,32 +40,40 @@ brew link --overwrite emacs-mac
 echo "linked emacs"
 
 echo "installing spacemacs"
-# git clone https://github.com/syl20bnr/spacemacs ~/.emacs.d || true
-git clone https://github.com/syl20bnr/spacemacs ~/.spacemacs.d || true
+distro='doom'
+# distro='spacemacs'
+# distro='vanilla'
 
-mkdir -p ~/.vanilla-emacs.d
-
-# I want to maintain a vanilla Emacs configuration alongside Spacemacs.
-# This is not fully baked. Spacemacs does not like seeing itself in a symlink,
-# and will add a lot of startup cost trying to deal with "duplicate packages".
-#
-# ln -snf ~/.vanilla-emacs.d ~/.emacs.d
-# ln -snf ~/.spacemacs.d ~/.emacs.d
-ln -snf ~/.emacs.d ~/.spacemacs.d
-
-echo "linking elisp dir"
-rm ~/.spacemacs.d/private/local/dotfiles || true
-ln -s -n -f $PWD/lisp ~/.spacemacs.d/private/local/dotfiles
-ln -snf $PWD/lisp/vanilla-init.el ~/.vanilla-emacs.d/init.el
-rm -rf ~/.spacemacs.d/private/snippets
-ln -s -n -f $PWD/yasnippets ~/.spacemacs.d/private/snippets
-# ln -s -n -f $PWD/emacs-config.org ~/.emacs.d/emacs-config.org
-echo "This directory managed by dotfiles" > ~/.spacemacs.d/private/local/dotfiles/README.org
-
-echo "Setting up additional layers"
-
-mkdir -p ~/.spacemacs.d/private/layers
-cd ~/.spacemacs.d/private/layers
+if [[ "$distro" == 'spacemacs' ]]; then
+  # I want to maintain a vanilla Emacs configuration alongside Spacemacs.
+  # This is not fully baked. Spacemacs does not like seeing itself in a symlink,
+  # and will add a lot of startup cost trying to deal with "duplicate packages".
+  #
+  git clone https://github.com/syl20bnr/spacemacs ~/.spacemacs.d || true
+  ln -snf ~/.spacemacs.d ~/.emacs.d
+  echo "linking elisp dir"
+  rm ~/.spacemacs.d/private/local/dotfiles || true
+  ln -s -n -f $PWD/lisp ~/.spacemacs.d/private/local/dotfiles
+  rm -rf ~/.spacemacs.d/private/snippets
+  ln -s -n -f $PWD/yasnippets ~/.spacemacs.d/private/snippets
+  # ln -s -n -f $PWD/emacs-config.org ~/.emacs.d/emacs-config.org
+  echo "This directory managed by dotfiles" > ~/.spacemacs.d/private/local/dotfiles/README.org
+  echo "Setting up additional layers"
+  mkdir -p ~/.spacemacs.d/private/layers
+  cd ~/.spacemacs.d/private/layers
+elif [[ "$distro" == 'doom' ]]; then
+  git clone --depth 1 https://github.com/hlissner/doom-emacs ~/.doom-emacs.d || true
+  ln -snf ~/.doom-emacs.d ~/.emacs.d
+  ~/.doom-emacs.d/bin/doom install
+  ln -sfn $PWD/emacs.d/early-init.el ~/.doom-emacs.d/early-init.el
+  ln -sfn $PWD/emacs.d/init.el ~/.doom-emacs.d/init.el
+  ln -sfn $PWD/doom/init.el ~/.doom.d/init.el
+  ln -sfn $PWD/doom/config.el ~/.doom.d/config.el
+  ln -sfn $PWD/doom/packages.el ~/.doom.d/packages.el
+elif [[ "$distro" == 'vanilla' ]]; then
+  mkdir -p ~/.vanilla-emacs.d
+  ln -snf $PWD/lisp/vanilla-init.el ~/.vanilla-emacs.d/init.el
+fi
 
 # we can preload the packages with this:
 # emacs --batch --load=~/.emacs.d/init.el
